backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.13
pydantic[email]==2.5.0
google-auth==2.23.4
PyJWT==2.8.0
python-dotenv==1.0.0